visualization func util.py 로 이동
